+++ ui/app/src/components/DialogHeader/DialogHeader.tsx
@@ -16,7 +16,6 @@
 
 import Typography, { TypographyProps } from '@mui/material/Typography';
 import IconButton from '@mui/material/IconButton';
-import { makeStyles } from 'tss-react/mui';
 import CloseIconRounded from '@mui/icons-material/CloseRounded';
 import MinimizeIconRounded from '@mui/icons-material/RemoveRounded';
 import ArrowBack from '@mui/icons-material/ArrowBackIosRounded';
@@ -30,46 +29,6 @@ import { SystemIconDescriptor } from '../SystemIcon';
 import { CSSObject } from 'tss-react';
 import Box from '@mui/material/Box';
 import { PartialSxRecord } from '../../models';
-
-const dialogTitleStyles = makeStyles()((theme) => ({
-  root: {
-    margin: 0,
-    display: 'flex',
-    flex: '0 0 auto',
-    flexWrap: 'wrap',
-    borderBottom: `1px solid ${theme.palette.divider}`,
-    padding: theme.spacing(1),
-    background: theme.palette.background.paper,
-    ...(theme.mixins.toolbar as CSSObject)
-  },
-  titleWrapper: {
-    display: 'flex',
-    width: '100%',
-    alignItems: 'center'
-  },
-  title: {
-    padding: `0 ${theme.spacing(1)}`,
-    overflow: 'hidden',
-    whiteSpace: 'nowrap',
-    textOverflow: 'ellipsis'
-  },
-  subtitle: {
-    fontSize: '14px',
-    lineHeight: '18px'
-  },
-  subtitleWrapper: {
-    padding: theme.spacing(1),
-    paddingTop: 0
-  },
-  leftActions: {
-    whiteSpace: 'nowrap'
-  },
-  rightActions: {
-    marginLeft: 'auto',
-    whiteSpace: 'nowrap'
-  },
-  backIcon: {}
-}));
 
 export interface DialogHeaderStateAction {
   icon: SystemIconDescriptor;
@@ -112,7 +71,6 @@ export type DialogHeaderProps<
   minimizeIcon?: ElementType;
   fullScreenIcon?: ElementType;
   backIcon?: ElementType;
-  classes?: Partial<Record<'root' | 'titleWrapper' | 'subtitleWrapper', string>>;
   className?: string;
   sxs?: PartialSxRecord<
     'root' | 'titleWrapper' | 'title' | 'subtitle' | 'subtitleWrapper' | 'leftActions' | 'rightActions' | 'backIcon'
@@ -126,7 +84,6 @@ export type DialogHeaderProps<
 
 export function DialogHeader(props: DialogHeaderProps) {
   // region
-  const { classes, cx } = dialogTitleStyles();
   const { formatMessage } = useIntl();
   const {
     id,
@@ -157,20 +114,27 @@ export function DialogHeader(props: DialogHeaderProps) {
   } = props;
   // endregion
   return (
-    <Box id={id} className={cx(className, classes.root, props.classes?.root)} sx={sxs?.root}>
-      <Box component="section" className={cx(classes.titleWrapper, props.classes?.titleWrapper)} sx={sxs?.titleWrapper}>
+    <Box
+      id={id}
+      className={className}
+      sx={{
+        margin: 0,
+        display: 'flex',
+        flex: '0 0 auto',
+        flexWrap: 'wrap',
+        borderBottom: (theme) => `1px solid ${theme.palette.divider}`,
+        padding: (theme) => theme.spacing(1),
+        background: (theme) => theme.palette.background.paper,
+        ...(theme) => theme.mixins.toolbar as CSSObject,
+        ...sxs?.root
+      }}
+    >
+      <Box component="section" sx={{ display: 'flex', width: '100%', alignItems: 'center', ...sxs?.titleWrapper }}>
         {(leftActions || onBack) && (
-          <Box className={classes.leftActions} sx={sxs?.leftActions}>
+          <Box sx={{ whiteSpace: 'nowrap', ...sxs?.leftActions }}>
             {onBack && (
               <Tooltip title={disabled ? '' : formatMessage(translations.back)}>
-                <IconButton
-                  aria-label="close"
-                  onClick={onBack}
-                  className={classes.backIcon}
-                  sx={sxs?.backIcon}
-                  size="large"
-                  disabled={disabled}
-                >
+                <IconButton aria-label="close" onClick={onBack} sx={sxs?.backIcon} size="large" disabled={disabled}>
                   <BackIcon />
                 </IconButton>
               </Tooltip>
@@ -180,11 +144,20 @@ export function DialogHeader(props: DialogHeaderProps) {
             ))}
           </Box>
         )}
-        <Typography className={classes.title} {...titleTypographyProps} sx={sxs?.title}>
+        <Typography
+          {...titleTypographyProps}
+          sx={{
+            padding: (theme) => `0 ${theme.spacing(1)}`,
+            overflow: 'hidden',
+            whiteSpace: 'nowrap',
+            textOverflow: 'ellipsis',
+            ...sxs?.title
+          }}
+        >
           {title}
         </Typography>
         {(rightActions || onCloseButtonClick || onMinimizeButtonClick || onFullScreenButtonClick) && (
-          <Box className={classes.rightActions} sx={sxs?.rightActions}>
+          <Box sx={{ marginLeft: 'auto', whiteSpace: 'nowrap', ...sxs?.rightActions }}>
             {rightActions?.map(({ icon, 'aria-label': tooltip, ...rest }: DialogHeaderActionProps, i: number) => (
               <Action key={i} icon={icon} tooltip={tooltip} disabled={disabled} {...rest} />
             ))}
@@ -218,13 +191,9 @@ export function DialogHeader(props: DialogHeaderProps) {
         )}
       </Box>
       {(subtitle || children) && (
-        <Box
-          component="section"
-          className={cx(classes.subtitleWrapper, props.classes?.subtitleWrapper)}
-          sx={sxs?.subtitleWrapper}
-        >
+        <Box component="section" sx={{ padding: (theme) => theme.spacing(1), paddingTop: 0, ...sxs?.subtitleWrapper }}>
           {subtitle && (
-            <Typography className={classes.subtitle} {...subtitleTypographyProps} sx={sxs?.subtitle}>
+            <Typography {...subtitleTypographyProps} sx={{ fontSize: '14px', lineHeight: '18px', ...sxs?.subtitle }}>
               {subtitle}
             </Typography>
           )}