+++ ui/app/src/pages/Login.tsx
@@ -15,7 +15,6 @@
  */
 
 import React, { useMemo } from 'react';
-import { makeStyles } from 'tss-react/mui';
 import LoginView, { LoginViewProps } from '../components/LoginView/LoginView';
 import CrafterThemeProvider from '../components/CrafterThemeProvider';
 import I18nProvider from '../components/I18nProvider';
@@ -24,26 +23,9 @@ import { UNDEFINED } from '../utils/constants';
 import ErrorState from '../components/ErrorState';
 import Dialog from '@mui/material/Dialog';
 import DialogContent from '@mui/material/DialogContent';
-
-const useStyles = makeStyles()(() => ({
-  root: {
-    height: '100%',
-    background: 'url("/studio/static-assets/images/cogs.jpg") 0 0 no-repeat',
-    backgroundSize: 'cover'
-  },
-  video: {
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-    minWidth: '100%',
-    position: 'fixed',
-    minHeight: '100%'
-  }
-}));
+import Box from '@mui/material/Box';
 
 export default function Login(props: LoginViewProps) {
-  const { classes } = useStyles();
   const supportError = useMemo<string>(
     () =>
       'SharedWorker' in window
@@ -67,19 +49,34 @@ export default function Login(props: LoginViewProps) {
   return (
     <I18nProvider>
       <CrafterThemeProvider>
-        <div className={classes.root}>
+        <Box
+          sx={{
+            height: '100%',
+            background: 'url("/studio/static-assets/images/cogs.jpg") 0 0 no-repeat',
+            backgroundSize: 'cover'
+          }}
+        >
           {/*
           **************************************************************************
           TODO: To be enabled or removed depending on the background video decision.
           **************************************************************************
-          <video
+          <Box
+            component="video"
             loop
             muted
             ref={videoRef}
             // autoPlay
             id="loginVideo"
             preload="auto"
-            className={classes.video}
+            sx={{
+              top: 0,
+              left: 0,
+              right: 0,
+              bottom: 0,
+              minWidth: '100%',
+              position: 'fixed',
+              minHeight: '100%'
+            }}
             poster="/studio/static-assets/images/camera-moving-through-cogs.jpeg"
             src="/studio/static-assets/images/camera-moving-through-cogs.mp4"
           />
@@ -97,7 +94,7 @@ export default function Login(props: LoginViewProps) {
           ) : (
             <LoginView {...props} />
           )}
-        </div>
+        </Box>
         <GlobalStyles />
       </CrafterThemeProvider>
     </I18nProvider>